fix(Android): Fixed PlayIntegrityHelper, cloudProjectNumber to long

+++ android/src/main/kotlin/com/example/root_detection/PlayIntegrityHelper.kt
@@ -15,7 +15,7 @@ class PlayIntegrityHelper(context: Context) {
     ) {
         val request = IntegrityTokenRequest.builder()
             .setNonce(nonce)
-            .setCloudProjectNumber(cloudProjectNumber)
+            .setCloudProjectNumber(cloudProjectNumber.toLong())
             .build()
 
         integrityManager.requestIntegrityToken(request)
